requirements.txt
flask>=3.0
anthropic>=0.40
gunicorn>=21.0
flask-httpauth>=4.8
python-dotenv>=1.0
gevent>=24.0
